+++ src/main/java/com/medirec/medirec/Repositories/PatientContactRepository.java
@@ -4,6 +4,6 @@ import com.medirec.medirec.Models.PatientContact;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientContactRepository extends JpaRepository<PatientContact, Long>{
+public interface PatientContactRepository extends JpaRepository<PatientContact, Integer>{
     
 }
