+++ src/main/java/com/medirec/medirec/Repositories/SymptomRepository.java
@@ -4,6 +4,6 @@ import com.medirec.medirec.Models.Symptom;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SymptomRepository extends JpaRepository<Symptom, Long>{
+public interface SymptomRepository extends JpaRepository<Symptom, Integer>{
     
 }
